% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectSiggene.R
\name{SelectSiggene}
\alias{SelectSiggene}
\title{Identify signature genes for each immune cell type compared with cancer cells}
\usage{
SelectSiggene(cancer_exp, immune_exp, marker_geneset, ratio = 0.05)
}
\arguments{
\item{cancer_exp}{the gene expression data set of cancer cell lines. A matrix with row names as Entrez gene ID and columns as cancer cell samples.}

\item{immune_exp}{the gene expression data set of different types of immune cells. A matrix with row names as Entrez gene ID and columns as immune cell samples.}

\item{marker_geneset}{list of marker genes of different immune cell subsets collected from literature research and alive tools.}

\item{ratio}{}
}
\value{
list of signature genes of each immune cell type.
}
\description{
the function is used to curate a specific gene set from overexpressed marker genes as signature genes for each immune cell type.
}
