% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StablePairs.R
\name{StablePairs}
\alias{StablePairs}
\title{get stable pairs from cancer cells}
\usage{
StablePairs(cancer_exp, geneid, freq)
}
\arguments{
\item{cancer_exp}{a dataset which is used to screen the stable pairs.A matrix with row names as Entrez gene ID and columns as cancer cell samples.}

\item{geneid}{the rownames of cancer_exp.}

\item{freq}{a threshold value is used to stipulate that a gene pair maintain the circumstance of > the threshold so that the gene pair should be retained.The threshold should between 0-1, and the default is 0.99.}
}
\value{
a matrix of stable pairs from cancer cells data.
}
\description{
the function is used to get highly stable gene pairs from a series of cancer cells.
}
\examples{

}
