% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stimulatedmodel.R
\name{Stimulatedmodel}
\alias{Stimulatedmodel}
\title{get parameters for upregulated scores transformation}
\usage{
Stimulatedmodel(combat_edata, select_siggene, stablepairs)
}
\arguments{
\item{combat_edata}{gene expression data including immune cells,cancer cells,and normal tissues using ComBat to remove batch effects. A matrix with row names as Entrez gene ID and columns as samples.}

\item{select_siggene}{list of signature genes for each immune cell type.}

\item{stablepairs}{a matrix of stable pairs from cancer cells with two columns,the first column is the higher expression gene,the second column is the lower expression gene.}
}
\value{
a matrix of transformed parameters for each immune cell type.
}
\description{
the function is used to design a transformation pipeline to transform the cell infiltration scores to the estimating cell proportions.
}
\examples{

}
