% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TICPE.R
\name{TICPEScores}
\alias{TICPEScores}
\title{Estimate the proportion of tumor-infiltrating immune cells}
\usage{
TICPEScores(expr, select_siggene, stablepairs, parameter, alpha = 0.5)
}
\arguments{
\item{expr}{the gene expression data set. A matrix with row names as Entrez gene ID and columns as samples.}

\item{select_siggene}{list of signature genes of each immune cell type.}

\item{stablepairs}{a matrix of stable pairs from cancer cells with two columns,the first column is the higher expression gene,the second column is the lower expression gene.}

\item{parameter}{a matrix of transformed parameters for each immune cell type.}

\item{alpha}{alpha a value to avoid division by 0. Default = 0.5.}
}
\value{
a matrix of the estimated proportion of tumor-infiltrating immune cells.
}
\description{
This package is used to estimate the proportion of tumor-infiltrating immune cells in a specific types of cancer.
}
\examples{

}
